--- src/navigation/components/index.tsx
@@ -1,1 +1,0 @@
-export { NavigationProvider, NavigationContext } from './NavigationProvider'
