+++ src/navigation/lib/initializeNavigationState.tsx
